refactor: rename components to kebab-casse
