+++ src/components/card-banner/card-banner-general.sass
