+++ src/components/card-banner/card-banner-resp.sass
