+++ src/components/card-hint/card-hint-general.sass
