+++ src/components/card-item/card-item-general.sass
