+++ src/components/card-item/card-item-resp.sass
