+++ src/components/card-list/card-list-general.sass
