+++ src/components/card-sub-banner/card-sub-banner-general.sass
