+++ src/components/card-sub-banner/card-sub-banner-resp.sass
