+++ src/components/card-timer/card-timer-general.sass
