+++ task-work.md
@@ -2,8 +2,6 @@
 
    - Удели внимание доступности для разных пользователей:
      - Используй атрибуты ARIA.
-     - Обеспечь возможность навигации и использования функционала с клавиатуры.
-     - Убедись, что контент корректно воспринимается скринридерами для незрячих пользователей.
 
 8. **Качество кода и оптимизация:**
